[MARTIFACT-80] Use per-module ignore configuration when generating aggregated buildinfo in `compare`
Use the per-module ignore configuration when generating buildinfo for comparison in the aggregator (usually the 'last' module). It's a bit of a bodge because it relies on the ignore config being stored when the `compare` goal runs in each module. If for some reason a module doesn't run the `compare` goal then it falls back to whatever config exists in the aggregator.

There's probably a much more complex solution that involves parsing the config out of the project model that would be more consistent.